Bytt ut SecureLog med TeamLog
Team Logs vil erstatte Securelog og pusher logger med
sensitive opplysninger til "team logs" i GCP i stedet
for til en tilgangsstyrt index i Elasticsearch.
