+++ nais/dev-fss.yaml
@@ -58,8 +58,6 @@ spec:
     type: RollingUpdate
   tokenx:
     enabled: true
-  secureLogs:
-    enabled: true
   prometheus:
     enabled: true
     path: /tiltaksgjennomforing-api/internal/actuator/prometheus
