+++ src/main/resources/logback-spring.xml
@@ -3,35 +3,32 @@
         <encoder class="net.logstash.logback.encoder.LogstashEncoder"/>
         <filter class="ch.qos.logback.core.filter.EvaluatorFilter">
             <evaluator class="ch.qos.logback.classic.boolex.OnMarkerEvaluator">
-                <marker>SECURE_LOG</marker>
+                <marker>TEAM_LOGS</marker>
             </evaluator>
-            <OnMismatch>NEUTRAL</OnMismatch>
             <OnMatch>DENY</OnMatch>
+            <OnMismatch>ACCEPT</OnMismatch>
         </filter>
     </appender>
 
-    <appender name="secureLog" class="ch.qos.logback.core.rolling.RollingFileAppender">
-        <file>/secure-logs/secure.log</file>
-        <rollingPolicy class="ch.qos.logback.core.rolling.FixedWindowRollingPolicy">
-            <fileNamePattern>/secure-logs/secure.log.%i</fileNamePattern>
-            <minIndex>1</minIndex>
-            <maxIndex>1</maxIndex>
-        </rollingPolicy>
-        <triggeringPolicy class="ch.qos.logback.core.rolling.SizeBasedTriggeringPolicy">
-            <maxFileSize>128MB</maxFileSize>
-        </triggeringPolicy>
-        <encoder class="net.logstash.logback.encoder.LogstashEncoder"/>
+    <appender name="team-logs" class="net.logstash.logback.appender.LogstashTcpSocketAppender">
+        <destination>team-logs.nais-system:5170</destination>
+        <encoder class="net.logstash.logback.encoder.LogstashEncoder">
+            <customFields>
+                {"google_cloud_project":"${GOOGLE_CLOUD_PROJECT}","nais_namespace_name":"${NAIS_NAMESPACE}","nais_pod_name":"${HOSTNAME}","nais_container_name":"${NAIS_APP_NAME}"}
+            </customFields>
+            <includeContext>false</includeContext>
+        </encoder>
         <filter class="ch.qos.logback.core.filter.EvaluatorFilter">
             <evaluator class="ch.qos.logback.classic.boolex.OnMarkerEvaluator">
-                <marker>SECURE_LOG</marker>
+                <marker>TEAM_LOGS</marker>
             </evaluator>
+            <OnMatch>ACCEPT</OnMatch>
             <OnMismatch>DENY</OnMismatch>
-            <OnMatch>NEUTRAL</OnMatch>
         </filter>
     </appender>
     <logger name="no.nav.vault.jdbc.hikaricp.VaultUtil" level="WARN"/>
     <root level="INFO">
         <appender-ref ref="consoleAppender"/>
-        <appender-ref ref="secureLog"/>
+        <appender-ref ref="team-logs"/>
     </root>
 </configuration>
